edison-oauth/src/main/java/de/otto/edison/oauth/ZonedDateTimeDeserializer.java
package de.otto.edison.oauth;

import com.fasterxml.jackson.core.JsonParser;
import com.fasterxml.jackson.databind.DeserializationContext;
import com.fasterxml.jackson.databind.JsonDeserializer;

import java.io.IOException;
import java.time.ZonedDateTime;

public class ZonedDateTimeDeserializer extends JsonDeserializer<ZonedDateTime> {
    @Override
    public ZonedDateTime deserialize(final JsonParser p, final DeserializationContext ctxt) throws IOException {
        return ZonedDateTime.parse(p.getText());
    }
}
